chore: add missing copyright and version details to third-party softwares file
A THIRD-PARTY must or should contain for each thirs-party software versions, copyrights, licenses
and URL to repository.
Such conditions are required for some licenses (like Apache 2.0) and mainly for distributed components
(e.g. those for app and lib parts).
Because we enjoy using third-party tools, even in the project, we should list them and credit their authors.

In few words, follow best practices, be compliant to compliance.
It was missing.

Signed-off-by: Pierre-Yves Lapersonne <pierreyves.lapersonne@orange.com>